* Fix windows build. * Fix windows x64 build. * Fix scripts build for x64. * Fix several annoying warnings on windows. * Add logging code for GPS. (will be useful when scripting) * Remove BT spells script from project and setup files, as it is deleted.

+++ src/ascnhalf-logonserver/AuthSocket.cpp
@@ -632,7 +632,7 @@ void AuthSocket::HandleReconnectChallenge()
     rs.SetRand(16*8);
     pkt.append(rs.AsByteBuffer());	// 16 bytes random
     pkt << (uint64) 0x00 << (uint64) 0x00;	// 16 bytes zeros
-    Send(pkt.contents(), pkt.size());
+    Send(pkt.contents(), (uint32)pkt.size());
 }
 
 void AuthSocket::HandleReconnectProof()
@@ -681,7 +681,7 @@ void AuthSocket::HandleReconnectProof()
 		pkt << (uint8)  0x03;	//ReconnectProof
 		pkt << (uint8)  0x00;
 		pkt << (uint16) 0x00;	// 2 bytes zeros
-		Send(pkt.contents(), pkt.size());
+		Send(pkt.contents(), (uint32)pkt.size());
 
 		// we're authenticated now :)
 		m_authenticated = true;
