+++ src/ascnhalf-realmserver/LogonCommClient.cpp
@@ -310,7 +310,7 @@ void LogonCommClientSocket::HandleRequestAccountMapping(WorldPacket & recvData)
 
 	ByteBuffer uncompressed(40000 * 5 + 8);
 	//uint32 Count = 0;
-	uint32 Remaining = mapping_to_send.size();
+	uint32 Remaining = (uint32)mapping_to_send.size();
 	itr = mapping_to_send.begin();
 	for(;;)
 	{
